requirements.txt
django==5.1.5
psycopg2-binary==2.9.10
djangorestframework==3.15.2
markdown==3.7
dj_database_url==2.3.0
whitenoise==6.8.2
gunicorn